requirements.txt
python==3.10
numpy==1.26.2
torch==2.1.1+cu121
pyro-ppl==1.8.6
git+https://github.com/kosiokarchev/clipppy.git
